pyproject.toml
[project]
name = "ezyassist-unified"
version = "1.0.0"
description = "EzyAssist Unified System - Telegram Bot + Registration Form"
authors = ["EzyAssist Team"]
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==20.7",
    "openai>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.10",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.2.0",
    "phonenumbers>=9.0.10",
    "pyjwt>=2.10.1"
]